chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [
    { name = "Subhashis Pradhan", email = "pradhan.subhasis174@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain-core (>=0.3.66,<0.4.0)",
    "langgraph (>=0.4.10,<0.5.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langgraph-checkpoint-postgres (>=2.0.21,<3.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
]

[tool.poetry]
packages = [{ include = "chatbot", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
